feat: move TimedCommitStatus into own folder, improve progress bar
Group TimedCommitStatus.tsx and its test under
plugins/TimedCommitStatus/ instead of flat in plugins/, matching the
one-component-per-folder layout the plugin architecture is heading
toward as more commit-status plugins are added.

Also fix the progress bar rendering as near-invisible text
highlighting: move the fill out from behind the label into its own
track below it, with a solid fill color instead of a translucent tint,
so the countdown reads as a progress bar rather than text selection.
