+++ ui/shared/src/components/plugins/TimedCommitStatus/TimedCommitStatus.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
-import type { Check, CommitStatusManager } from '../../types/promotion';
-import type { components } from '../../types/generated/view.gen';
-import { formatDuration } from '../../utils/util';
+import type { Check, CommitStatusManager } from '../../../types/promotion';
+import type { components } from '../../../types/generated/view.gen';
+import { formatDuration } from '../../../utils/util';
 
 export interface TimedCommitStatusProps {
   check: Check;
@@ -79,22 +79,35 @@ const TimedCommitStatus: React.FC<TimedCommitStatusProps> = ({ check, manager })
     requiredDurationMs > 0 ? Math.min(Math.max(clampedRemaining / requiredDurationMs, 0), 1) : 0;
 
   return (
-    <div style={{ position: 'relative', overflow: 'hidden' }}>
-      <div
-        className="timed-commit-status-fill"
-        style={{
-          position: 'absolute',
-          top: 0,
-          left: 0,
-          bottom: 0,
-          width: `${ratio * 100}%`,
-          backgroundColor: 'rgba(66, 133, 244, 0.2)',
-          transition: 'width 1s linear',
-        }}
-      />
-      <span style={{ position: 'relative' }} className="check-name-text">
+    <div>
+      <span className="check-name-text">
         {check.name} ({formatDuration(clampedRemaining)} remaining)
       </span>
+      <div
+        className="timed-commit-status-track"
+        style={{
+          position: 'relative',
+          height: 6,
+          marginTop: 4,
+          borderRadius: 3,
+          backgroundColor: 'rgba(66, 133, 244, 0.15)',
+          overflow: 'hidden',
+        }}
+      >
+        <div
+          className="timed-commit-status-fill"
+          style={{
+            position: 'absolute',
+            top: 0,
+            left: 0,
+            bottom: 0,
+            width: `${ratio * 100}%`,
+            borderRadius: 3,
+            backgroundColor: 'rgb(66, 133, 244)',
+            transition: 'width 1s linear',
+          }}
+        />
+      </div>
     </div>
   );
 };
