+++ ui/shared/src/components/plugins/index.ts
@@ -1,6 +1,6 @@
 import type React from 'react';
 import type { Check, CommitStatusManager, CommitStatusManagerKind } from '../../types/promotion';
-import TimedCommitStatus from './TimedCommitStatus';
+import TimedCommitStatus from './TimedCommitStatus/TimedCommitStatus';
 
 export const commitStatusPlugins: Partial<
   Record<CommitStatusManagerKind, React.FC<{ check: Check; manager: CommitStatusManager }>>
